fix: amount 소수점 허용 (USD 9.72 같은 값 저장 가능)
- TradeInput Pydantic 모델: amount int → float
- 프론트: Math.round 제거, step="any" 로 소수점 입력 허용

+++ api.py
@@ -760,7 +760,7 @@ class TradeInput(BaseModel):
     section_name: str
     month: str
     stock_name: str
-    amount: int
+    amount: float   # USD는 소수점 허용 (e.g. 9.72)
     realized: bool = False
